feat: store arrival and departure time as primitive ints and create service daytime objects on the fly. Change transfer time from int to short.

+++ libs/gtfs/src/main/java/org/naviqore/gtfs/schedule/model/Transfer.java
@@ -17,9 +17,9 @@ public class Transfer {
     private final Stop toStop;
     private final TransferType transferType;
     @Nullable
-    private final Integer minTransferTime;
+    private final Short minTransferTime;
 
-    public Optional<Integer> getMinTransferTime() {
+    public Optional<Short> getMinTransferTime() {
         return Optional.ofNullable(minTransferTime);
     }
